support study-level matrix route when subject is unspecified
